requirements.txt
nonebot2>=2.4.0,<3.0.0
matplotlib>=3.0.0
meme-generator>=0.0.18,<1.0.0
Pillow
aiohttp>=3.11.10,<4.0.0
deepl>=1.20.0,<2.0.0
python-socketio>=5.11.4,<6.0.0
qrcode>=8.0,<9.0
SQLAlchemy>=2.0.36,<3.0.0
playwright>=1.49.1,<2.0.0
beautifulsoup4>=4.12.3,<5.0.0
toml>=0.10.2,<1.0.0
ruff>=0.8.3,<1.0.0
aiofiles>=24.1.0,<25.0.0
python-multipart>=0.0.6,<1.0.0
pytest>=8.3.4,<9.0.0
pytest-asyncio>=0.25.0,<1.0.0
websockets>=14.0.0,<15.0.0
httpx ==0.27.2
arclet-alconna >=1.8.0
qrcode>=8.0,<9.0
openai

nb-cli>=1.4.2,<2.0.0
nonebot2[fastapi]>=2.4.0,<3.0.0

nonebot-plugin-alconna >=0.5.0
nonebot_plugin_waiter >=0.8.0
nonebot_plugin_apscheduler>=0.5.0,<1.0.0
nonebot_plugin_memes>=0.6.2,<1.0.0
nonebot_plugin_userinfo

nonebot-adapter-onebot>=2.4.6,<3.0.0
nonebot-adapter-kaiheila
nonebot-adapter-discord